requirements.txt
fastapi
motor
uvicorn
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
awesome-slugify